+++ morpheus-graphql/src/Data/Morpheus/Server/Types/Types.hs
@@ -22,11 +22,7 @@ import Prelude
 
 data Undefined (m :: Type -> Type) = Undefined deriving (Show, Generic)
 
-data Pair k v = Pair
-  { key :: k,
-    value :: v
-  }
-  deriving (Generic)
+data Pair k v = Pair k v deriving (Generic)
 
 data TypeGuard interface union
   = ResolveInterface interface